organize integration services
Move MQTT, weather, notifications, email, SMTP, and ESP network setup under services/. Keep forwarding headers at the repository root for source compatibility and update native and PlatformIO source discovery for the shallow module layout.
